+++ Changelog.md
@@ -1,9 +1,13 @@
 # Change Log
 
-## [0.xx.x] - Unreleased
+## [0.14.2] - Unreleased
 
 ### Added
   - Syntax: register/field array index now support comma separated and/or range index (like `[0:3,7]`)
+  - Generics: support generics for optional register instance (not inside expression)
+
+### Fixed
+  - Generics: the address error signal was inverted
 
 ## [0.14.1] - 2025-07-18